LLM_practice01/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
openai==0.28.0
propcache==0.2.1
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
yarl==1.18.3
